The code implementing the conditions is written below:
Split a search query into individual words and form a list of unique lemmas from these words, excluding interjections, conjunctions, prepositions and particles. To do this, use the code that you already wrote in the previous step.

Exclude lemmas that occur on too many pages from the resulting list. Experiment and determine this percentage yourself.

Sort the lemmas in order of increasing frequency (in ascending order of the value of the frequency field) — from the rarest to the most frequent.

Using the first, rarest lemma from the list, find all the pages on which it occurs. Next, look for the matches of the next lemma from this list of pages, and then repeat the operation for each subsequent lemma. The list of pages should decrease at each iteration.

If there are no pages left in the end, then display an empty list.