If the pages are found, calculate the relevance for each of them.
For each page, calculate the absolute relevance — the sum of all the lemmas found on the page (from the index table), which is divided by the maximum value of this absolute relevance for all the pages found.

Relative relevance can be obtained by dividing the absolute relevance for a particular page by the maximum absolute relevance among all pages of a given search output.